Fixed paths after modules moving
